\name{aispu-package}
\alias{aispu-package}
\docType{package}
\title{
\packageTitle{aispu}
}
\description{
\packageDescription{aispu}
}
\details{
This package provides a new test called the adaptive interaction sum of powered score (aiSPU) test based on the truncated Lasso penalty and an adaptive testing idea. Some related methods have been implemented in this package as well. 
}
\author{
\packageAuthor{aispu}

Maintainer: \packageMaintainer{aispu}
}
\references{
Wu, C., Xu, G., Shen, X., & Pan, W. (2020+). A Regularization-Based Adaptive Test for High-Dimensional Generalized Linear Models, Submitted.

Zhang, X., & Cheng, G. (2017). Simultaneous inference for high-dimensional linear models. Journal of the American Statistical Association, 112(518), 757-768.
}

\keyword{ package }

