% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aispu.R
\name{aispu}
\alias{aispu}
\title{Adaptive Interaction Sum of Powered Score (aiSPU) Tests}
\usage{
aispu(Y, X, cov = NULL, cov2, pow = c(1:6, Inf), model = c("gaussian",
  "binomial"), n.perm = 1000, penalty = c("tlp", "lasso", "ridge", "net",
  "mcp", "SCAD"), tau = 0.1, standardize = FALSE, dfmax = 1000,
  pmax = 1000, resample = c("asy", "boot"))
}
\arguments{
\item{Y}{Response or phenotype data. It can be either a binary trait or a quantitative trait. A vector with length n (number of subjects).}

\item{X}{Variables of interest; each row for a subject, and each column
for predictor).A matrix with dimension n by p.}

\item{cov}{Ancillary covariates without penalization, such as age, gender, and etc.}

\item{cov2}{High-dimensional covariates with penalization.}

\item{pow}{Power used in iSPU test. A vector of the powers.}

\item{model}{Use "gaussian" for a quantitative trait, and use "binomial" for a binary trait.}

\item{n.perm}{The number of boostrap replications}

\item{penalty}{Penalty for cov2. The default is truncated lasso (tlp). We recommend using tlp in real data applications.}

\item{tau}{Tunning parameters for tlp.}

\item{standardize}{Logical flag for x variable standardization, prior to fitting the model sequence. The coefficients are always returned on the original scale. Default is standardize=TRUE. If variables are in the same units already, you might not wish to standardize.}

\item{dfmax}{Limit the maximum number of variables in the model. Useful when the dimension of cov2 is high.}

\item{pmax}{Limit the maximum number of variables ever to be nonzero.}

\item{resample}{Methods for calculating p-values. The default is the asymptotics-based method ("asy").}
}
\value{
A list object, Ts : test statistics for the iSPU tests (in the order of the specified pow) and finally for the aSPU test.
        pvs : p-values for the iSPU and aiSPU tests.
}
\description{
It returns p-values of the iSPU tests and aiSPU test.
}
\examples{

# Generate the data (codes for the simulations in the manuscript)
n = 30
signal.r = 0
nInformative = 3
p = 40
seed = 1
s = 0.01
non.zero = floor((p/2) * s)
alpha = c(rep(0,p/2 - non.zero), runif(non.zero,-signal.r,signal.r))
beta = c(rep(2,nInformative), rep(0,(p/2- 3)), alpha)
dat = sim_data(seed, n = n, p = p, beta = beta)

X = dat$X
Y = dat$Y

cov = NULL
X.tmp = X
cov2 = X.tmp[,1:(p/2)]
X = X.tmp[,(p/2 + 1):p]
aispu(Y, X,cov = NULL, cov2, pow = c(1:6, Inf), model= "gaussian",penalty = "tlp", n.perm = 10,resample = "boot")
}
\references{
Wu, C., Xu, G., Shen, X., & Pan, W. (2018+). An adaptive test for high-dimensional generalized linear models with application to detect gene-environment interactions, Submitted.
}
\author{
Chong Wu and Wei Pan
}
