% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Generate Simulated Data}
\usage{
sim_data(seed, n, p, beta)
}
\arguments{
\item{seed}{Random seed.}

\item{n}{Number of subjects.}

\item{p}{Number of variables.}

\item{beta}{Coefficients.}
}
\value{
A list object containing the simulated data.
}
\description{
It returns simulated data. This is corresponding to high dimensional linear model simulation settings in the manuscript.
}
\examples{

# Generate the data (codes for the simulations in the manuscript)
n = 30
signal.r = 0
nInformative = 3
p = 40
seed = 1
s = 0.01
non.zero = floor((p/2) * s)
alpha = c(rep(0,p/2 - non.zero), runif(non.zero,-signal.r,signal.r))
beta = c(rep(2,nInformative), rep(0,(p/2- 3)), alpha)
dat = sim_data(seed, n = n, p = p, beta = beta)

X = dat$X
Y = dat$Y

}
\references{
Wu, C., Xu, G., Shen, X., & Pan, W. (2018+). An adaptive test for high-dimensional generalized linear models with application to detect gene-environment interactions, Submitted.
}
\author{
Chong Wu and Wei Pan
}
